Feacture/controller impedimento

+++ src/main/java/br/com/zup/gerenciador/de/dailys/impedimento/ImpedimentoService.java
@@ -1,7 +1,5 @@
 package br.com.zup.gerenciador.de.dailys.impedimento;
 
-import br.com.zup.gerenciador.de.dailys.impedimento.Impedimento;
-import br.com.zup.gerenciador.de.dailys.impedimento.ImpedimentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,6 +21,11 @@ public class ImpedimentoService {
     }
 
     public void deletarImpedimento(Long id){
+        if(!impedimentoRepository.existsById(id)){
+            throw new ImpedimentoInexistente("Impedimento não encontrado");
+
+        }
         impedimentoRepository.deleteById(id);
     }
+
 }
